Add files via upload
Communications have evolved over the past years allowing people to constantly interact with others 24/7 no matter where you or the message/call receiver is located. Customers push Telecomunication organizations to compete between them to offer the best price and/or quality service. Hence, Telco companies must understand what drives customer churn and how to deal with it.

As a business analyst of a Communication Company, I will create models to predict the likelihood of churning of a particular client based on significant attributes that lead him/her to make the decision to go somewhere else. Furthermore, I will measure the financial implication of losing an existing client using an average of currently offered cellular plans to calculate the drop in sales and profit.